create new endpoints

+++ Catalog.Application/DTOs/ContentDto.cs
@@ -4,4 +4,4 @@ public record ContentDto(
     Guid Id,
     string Name,
     string Description,
-    int Sla);
+    double Sla);
