+++ Catalog.Application/DTOs/SpecificCategoryDto.cs
@@ -4,4 +4,4 @@ public record SpecificCategoryDto(
     Guid Id,
     string Name,
     string Description,
-    int Sla);
+    double Sla);
