+++ Catalog.Application/UseCases/Category/Create/CategoryCreateRequest.cs
@@ -6,4 +6,4 @@ namespace Catalog.Application.UseCases.Category.Create;
 
 public record CategoryCreateRequest(
     string Name,
-    double Sla): IRequest<Response>;
+    int Sla): IRequest<Response>;
