+++ Catalog.Application/UseCases/Category/Update/CategoryUpdateRequest.cs
@@ -6,4 +6,4 @@ namespace Catalog.Application.UseCases.Category.Update;
 public record CategoryUpdateRequest(
     Guid Id,
     string Name,
-    double Sla): IRequest<Response>;
+    int Sla): IRequest<Response>;
